New styled calendar in goals, and styling in goals and agenda. Also setted max and min in calendars
